Register an IPython magic directly
- Follow-up to https://github.com/yuvipanda/github-app-user-auth/pull/8,
  we now register an IPython magic directly when applicable so
  users can call %ghauth same as they would from the terminal.
- Provide slightly better error message when github app id is
  mis-specified
- Move everything into one file, as that's all the code we have.
  We can split up the file later if we add more code
