--- github_app_user_auth/__main__.py
@@ -1,3 +1,0 @@
-from .auth import main
-
-main()
