--- github_app_user_auth/jupyter.py
@@ -1,4 +1,0 @@
-if __name__ == '__main__':    
-    from .auth import main
-    
-    main(in_jupyter=True)
